+++ src/types/multiscan-step.tsx
@@ -1,9 +1,12 @@
-type TMultiscanStep = 'initial' |
+type TMultiscanStep = 'not_initialized' |
+                 'initial' |
                  'wallets_list' |
                  'download_await' |
                  'zerion_connection' |
                  'wallet_redirect_await' |
                  'set_address' |
-                 'crossmint_connection'
+                 'crossmint_connection' |
+                 'sign_message' |
+                 'eligible_to_claim'
 
 export default TMultiscanStep