+++ src/types/whitelist-type.tsx
@@ -1,0 +1,3 @@
+type TWhitelistType = 'address' | 'email' | 'twitter'
+
+export default TWhitelistType